chore: move TestnetBanner into parts
